refactor : 관리자 분리
