+++ .gitignore
@@ -26,7 +26,6 @@ out/
 !**/src/main/**/out/
 !**/src/test/**/out/
 .env
-application.yml
 
 ### NetBeans ###
 /nbproject/private/
