+++ run_core/src/main/java/hyper/run/domain/game/service/GameService.java
@@ -40,7 +40,7 @@ public class GameService {
     private final GameRepository gameRepository;
     private final GameHistoryRepository gameHistoryRepository;
     private final GameScheduler gameScheduler;
-    private final GameRepositoryCustom gameRepositoryCustom;
+
 
     public void testStart(Long gameId){
         Game game = gameRepository.findById(gameId).get();
@@ -171,16 +171,6 @@ public class GameService {
         }
         return GameStatus.REGISTRATION_OPEN;
     }
-    /** 관리자 페이지
-     * 예정,진행,종료된 경기 모두 조회
-     */
-    public Page<AdminGameResponse> findAllGames(LocalDate startDate, LocalDate endDate, AdminGameStatus status, String keyword, Pageable pageable) {
 
-        LocalDateTime createdAfter = (startDate != null) ? startDate.atStartOfDay() : null;
-        LocalDateTime createdBefore = (endDate != null) ? endDate.plusDays(1).atStartOfDay() : null;
-
-        Page<Game> games = gameRepositoryCustom.findGamesByCriteria(createdAfter, createdBefore, status, keyword, pageable);
-        return games.map(AdminGameResponse::dtoFromGame);
-    }
 
 }
