+++ run_core/src/main/java/hyper/run/domain/inquiry/dto/request/InquirySearchRequest.java
@@ -5,6 +5,7 @@ import hyper.run.domain.inquiry.entity.InquiryType;
 import lombok.*;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,8 +13,8 @@ import java.time.LocalDate;
 @Builder
 @Setter
 public class InquirySearchRequest {
-    private LocalDate startDate;
-    private LocalDate endDate;
+    private LocalDateTime startDate;
+    private LocalDateTime endDate;
     private String keyword;
     private InquiryState state;
     private InquiryType type;
